Fix default optimizer in FloodModelParams.
The default optimizer is being used across all instances of the dataclass
that don't specify a different value. Because this is currently a
keras.Optimizer object, that same object is getting passed to possibly
multiple models, which will throw an error.

Changed this to a string, which is also a valid optimizer value. As with
all other model parameters, we leave it to the user to supply reasonable
values.